Move evrything about configuration in the Configuration class
