+++ src/ansys/platform/instancemanagement/configuration.py
@@ -1,10 +1,16 @@
 """Configuration class module."""
 import json
 import logging
+import os
 import re
 from typing import Sequence, Tuple
 
-from ansys.platform.instancemanagement.exceptions import InvalidConfigurationError
+from ansys.platform.instancemanagement.exceptions import (
+    InvalidConfigurationError,
+    NotConfiguredError,
+)
+
+CONFIGURATION_PATH_ENVIRONMENT_VARIABLE = "ANSYS_PLATFORM_INSTANCEMANAGEMENT_CONFIG"
 
 logger = logging.getLogger(__name__)
 
@@ -137,3 +143,58 @@ Consider upgrading ansys-platform-instancemanagement.',
         else:
             access_token = None
         return Configuration(headers, tls, uri, access_token)
+
+    @staticmethod
+    def from_environment():
+        """Create a PyPIM configuration based on the environment configuration.
+
+        Before calling this method, :func:`~is_configured()` should be called to check if
+        the environment is configured to use PyPIM.
+
+        The environment configuration consists in setting the environment variable
+        ``ANSYS_PLATFORM_INSTANCEMANAGEMENT_CONFIG`` to the path of the PyPIM
+        configuration file. The configuration file is a simple JSON file containing
+        the URI of the PIM API and the headers required to pass information.
+
+        The configuration file format is:
+
+        .. code-block:: json
+
+            {
+                "version": 1,
+                "pim": {
+                    "uri": "dns:pim.svc.com:80",
+                    "headers": {
+                        "metadata-info": "value"
+                    },
+                    "tls": false
+                }
+            }
+
+        Returns
+        -------
+        Configuration
+            PyPIM configuration settings.
+
+        Raises
+        ------
+        NotConfiguredError
+            The environment is not configured to use PyPIM.
+
+        InvalidConfigurationError
+            The configuration is invalid.
+        """
+        if not is_configured():
+            raise NotConfiguredError("The environment is not configured to use PyPIM.")
+        return Configuration.from_file(os.environ[CONFIGURATION_PATH_ENVIRONMENT_VARIABLE])
+
+
+def is_configured() -> bool:
+    """Check if the environment is configured to use PyPIM.
+
+    Returns
+    -------
+    bool
+        ``True`` when the environment is configured to use PyPIM, ``False`` otherwise.
+    """
+    return CONFIGURATION_PATH_ENVIRONMENT_VARIABLE in os.environ
